datadog-reporter: add container name pprof label
permits slicing down the profile by container name.

we use container_name instead of containerName as we're currently
normalizing the tag to containername, and this leads to a
discrepancy between the tag and the pprof label (containerName).

also mount docker.sock locally so that the otel-agent is able
to find the container names