container metadata: add container_id as a label (PROF-9951) (#18)
* container metadata: add container_id as a label

add a default parsing pattern for the cgroup file based on https://github.com/open-telemetry/opentelemetry-go/blob/6a0fa3c911a061198302cd1972c2e1a9e5ffa87f/sdk/resource/container.go#L21
to be able to parse the container ID even when the container metadata
environment is not detected.

export containerID attribute to be able to inject it in the traceInfo

also hook this up as a pprof label, and a tag for the dd reporter.