pfelf: refactor debug links support
add OpenDebugBuildID method and refactor OpenDebugLink method so that
it's compatible with https://sourceware.org/gdb/current/onlinedocs/gdb.html/Separate-Debug-Files.html

this is the way that gdb searches for separate debug executables, so it
should hopefully be more exhaustive.

as a side note, -dbgsym packages on ubuntu (for example,
openssh-server-dbgsym) only download symbols to the build ID directory
/usr/lib/debug/.build-id/ (and not the debug link directories), hence
the new method also maximizes the chance to find potential debug symbols
on ubuntu.