v8: only remove prefix when mapping is removed
prior to this change we were tracking generation number (i.e.
count of SynchronizeMappings call where the object was found) for
both mappings and prefixes.

if the mapping did not change, we would not re-calculate prefixes
for the mapping, hence we would not increase the generation of the
prefix object.

as such, we would end up cleaning the prefix on the second call of
SynchronizeMappings (if mappings did not change), removing the prefix
from the ebpf map, and making unwinding fail for PC in the JIT area.

to fix this, we only track the generation for mappings. when mappings
are removed, we computed all the prefixes associated to those mappings
and we remove them. this should be safe because different mappings
should have different prefixes.